Parse and dispatch the 'list' subcommand

Assembles arguments from the parsed argparse result and calls cli_nemo_list(). When wf is non-NULL it overrides args$workflow (used by downstream packages that fix the workflow at the script level, e.g. tidywigits.R).

Usage

cli_list_parse_args(args, wf = NULL)

Arguments

args

Named list of parsed CLI arguments, as returned by argparse. Expected fields: format, in_dir, workflow (may be NULL when wf is provided), max, quiet.

wf

(character(1) or NULL)
Workflow override. When non-NULL, replaces args$workflow.

Examples

path <- system.file("extdata/tool1", package = "nemo")
args <- list(format = "pretty", in_dir = path, workflow = NULL, quiet = FALSE)
capture.output(cli_list_parse_args(args, wf = "workflow1"))
#>  [1] "|  n|tool_parser  |prefix    |bname                    | size|lastmodified        |path                                                                                                      |"
#>  [2] "|--:|:------------|:---------|:------------------------|----:|:-------------------|:---------------------------------------------------------------------------------------------------------|"
#>  [3] "|  1|tool1_table1 |sampleA   |sampleA.tool1.table1.tsv |   93|2026-06-04 15:05:16 |/home/runner/miniconda3/envs/pkgdown_env/lib/R/library/nemo/extdata/tool1/v4.5.6/sampleA.tool1.table1.tsv |"
#>  [4] "|  2|tool1_table1 |sampleA_2 |sampleA.tool1.table1.tsv |  113|2026-06-04 15:05:16 |/home/runner/miniconda3/envs/pkgdown_env/lib/R/library/nemo/extdata/tool1/v1.2.3/sampleA.tool1.table1.tsv |"
#>  [5] "|  3|tool1_table1 |sampleA_3 |sampleA.tool1.table1.tsv |  133|2026-06-04 15:05:16 |/home/runner/miniconda3/envs/pkgdown_env/lib/R/library/nemo/extdata/tool1/latest/sampleA.tool1.table1.tsv |"
#>  [6] "|  4|tool1_table2 |sampleA   |sampleA.tool1.table2.tsv |   47|2026-06-04 15:05:16 |/home/runner/miniconda3/envs/pkgdown_env/lib/R/library/nemo/extdata/tool1/v1.0.0/sampleA.tool1.table2.tsv |"
#>  [7] "|  5|tool1_table2 |sampleA_2 |sampleA.tool1.table2.tsv |   70|2026-06-04 15:05:16 |/home/runner/miniconda3/envs/pkgdown_env/lib/R/library/nemo/extdata/tool1/latest/sampleA.tool1.table2.tsv |"
#>  [8] "|  6|tool1_table3 |sampleA   |sampleA.tool1.table3.tsv |   48|2026-06-04 15:05:16 |/home/runner/miniconda3/envs/pkgdown_env/lib/R/library/nemo/extdata/tool1/v1.0.0/sampleA.tool1.table3.tsv |"
#>  [9] "|  7|tool1_table3 |sampleA_2 |sampleA.tool1.table3.tsv |   83|2026-06-04 15:05:16 |/home/runner/miniconda3/envs/pkgdown_env/lib/R/library/nemo/extdata/tool1/latest/sampleA.tool1.table3.tsv |"
#> [10] "|  8|tool1_table4 |sampleA   |sampleA.tool1.table4.tsv |   34|2026-06-04 15:05:16 |/home/runner/miniconda3/envs/pkgdown_env/lib/R/library/nemo/extdata/tool1/v1.0.0/sampleA.tool1.table4.tsv |"
#> [11] "|  9|tool1_table4 |sampleA_2 |sampleA.tool1.table4.tsv |   52|2026-06-04 15:05:16 |/home/runner/miniconda3/envs/pkgdown_env/lib/R/library/nemo/extdata/tool1/latest/sampleA.tool1.table4.tsv |"
#> [12] "| 10|tool1_table5 |sampleA   |sampleA.tool1.table5.csv |  994|2026-06-04 15:05:16 |/home/runner/miniconda3/envs/pkgdown_env/lib/R/library/nemo/extdata/tool1/latest/sampleA.tool1.table5.csv |"
#> [13] "| 11|tool1_table6 |sampleA   |sampleA.tool1.table6.csv |   78|2026-06-04 15:05:16 |/home/runner/miniconda3/envs/pkgdown_env/lib/R/library/nemo/extdata/tool1/v1.0.0/sampleA.tool1.table6.csv |"
#> [14] "| 12|tool1_table6 |sampleA_2 |sampleA.tool1.table6.csv |  100|2026-06-04 15:05:16 |/home/runner/miniconda3/envs/pkgdown_env/lib/R/library/nemo/extdata/tool1/latest/sampleA.tool1.table6.csv |"