Tool Object

Base R6 class for all nemo tools. Subclasses implement parsers for specific bioinformatic tool outputs by optionally overriding parse_{tname}() and tidy_{tname}() methods for custom parse or tidy logic per table.

A Tool object:

has a name (name);

has a path to a directory with its outputs (path);

has a schema configuration Config object (config);

has a tibble of files matching its Config patterns (files);

has a tibble with the parsed and tidied files (tbls);

The typical workflow is: optionally filter files with filter_files(), parse and tidy with tidy(), then write outputs with write(). nemofy() chains all three steps.

Public fields

name

(character(1))
Name of tool.

pkg

(character(1))
Package name tool belongs to (for config lookup).

path

(character(1))
Output directory of tool.

config

(Config())
Config of tool.

files

(tibble())
Tibble of files matching available Tool patterns.

tbls

(tibble())
Tibble of tidy tibbles.

raw_schemas_all

(tibble())
All raw schemas for tool.

tidy_schemas_all

(tibble())
All tidy schemas for tool.

get_tidy_schema

(function())
Get specific tidy schema.

get_raw_schema

(function())
Get specific raw schema.

get_col_map

(function())
Get column mapping (raw -> tidy) for a table.

files_tbl

(tibble(n))
Tibble of files from list_files_dir().

Methods

Public methods

Tool$new()

Tool$print()

Tool$filter_files()

Tool$list_files()

Tool$.dispatch_parse()

Tool$.dispatch_tidy()

Tool$.parse_by_ftype()

Tool$.parse_file()

Tool$.tidy_file()

Tool$.parse_file_keyvalue()

Tool$.parse_file_nohead()

Tool$tidy()

Tool$write()

Tool$nemofy()

Tool$clone()

Method new()

Create a new Tool object.

Usage

Tool$new(name = NULL, pkg = NULL, path = NULL, files_tbl = NULL)

Arguments

name

(character(1))
Name of tool.

pkg

(character(1))
Package name tool belongs to (for config lookup).

path

(character(1))
Output directory of tool. If files_tbl is supplied, this basically gets ignored.

files_tbl

(tibble(n))
Tibble of files from list_files_dir().

Returns

(R6::R6Class())
R6 object.

Method print()

Print details about the Tool.

Usage

Tool$print(...)

Arguments

...

(ignored).

Returns

(R6::R6Class())
R6 object invisibly.

Method filter_files()

Filter files in given tool directory based on inclusion or exclusion tool_parser names. The result is reflected in the files field.

Usage

Tool$filter_files(include = NULL, exclude = NULL)

Arguments

include

(character(n))
tool_parser names to include (e.g. "tool1_table1").

exclude

(character(n))
tool_parser names to exclude (e.g. "tool1_table3").

Returns

(R6::R6Class())
R6 object invisibly.

Method list_files()

List only files of interest in given tool directory, i.e. only those files that match the patterns listed in the tool config.

Usage

Tool$list_files(type = "file")

Arguments

type

(character(1))
File type(s) to return (e.g. any, file, directory, symlink). See fs::dir_info.

Returns

(tibble())
A tibble with:

tool_parser: tool name followed by parser name;

parser: parser name;

bname: file basename;

size: file size;

lastmodified: last modified timestamp of file;

path: file path;

pattern: file pattern;

prefix: file prefix;

group: if multiple files have the same basename then this is used as a differentiator.

Method .dispatch_parse()

Dispatch parse for a table: calls custom parse_{name}() if defined in the subclass, otherwise falls back to .parse_by_ftype().

Usage

Tool$.dispatch_parse(x, tname)

Arguments

x

(character(1))
File path.

tname

(character(1))
Table name.

Returns

(tibble())
Parsed data in tibble.

Method .dispatch_tidy()

Dispatch tidy for a table: calls custom tidy_{tname}() if defined in the subclass, otherwise falls back to .tidy_file().

Usage

Tool$.dispatch_tidy(x, tname)

Arguments

x

(character(1) or tibble())
File path or already parsed raw tibble.

tname

(character(1))
Table name.

Returns

(tibble())
Tidy data in enframed tibble.

Method .parse_by_ftype()

Parse a file by looking up its ftype from the config and calling the appropriate internal parser. Errors for ftypes that require a custom parse_{tname}() method (e.g. csv-nohead-long).

Usage

Tool$.parse_by_ftype(x, tname)

Arguments

x

(character(1))
File path.

tname

(character(1))

Returns

(tibble())
Parsed data in tibble.

Method .parse_file()

Parse file.

Usage

Tool$.parse_file(x, name, delim = "\t", ...)

Arguments

x

(character(1))
File path.

name

(character(1))
Parser name (e.g. "breakends" - see docs).

delim

(character(1))
File delimiter.

...

Passed on to readr::read_delim.

Returns

(tibble())
Parsed data in tibble.

Method .tidy_file()

Tidy file.

Usage

Tool$.tidy_file(x, name, convert_types = FALSE)

Arguments

x

(character(1) or tibble())
File path or already parsed raw tibble.

name

(character(1))
Parser name (e.g. "breakends" - see docs).

convert_types

(logical(1))
Convert field types based on schema.

Returns

(tibble())
Tidy data in enframed tibble.

Method .parse_file_keyvalue()

Parse files with no header and two columns representing key-value pairs.

Usage

Tool$.parse_file_keyvalue(x, name, delim = "\t", ...)

Arguments

x

(character(1))
File path.

name

(character(1))
Parser name (e.g. "qc" - see docs).

delim

(character(1))
File delimiter.

...

Passed on to readr::read_delim.

Returns

(tibble())
Parsed data in tibble.

Method .parse_file_nohead()

Parse headless file.

Usage

Tool$.parse_file_nohead(x, pname, delim = "\t", ...)

Arguments

x

(character(1))
File path.

pname

(character(1))
Parser name (e.g. "breakends" - see docs).

delim

(character(1))
File delimiter.

...

Passed on to readr::read_delim.

Returns

(tibble())
Parsed data in tibble.

Method tidy()

Tidy a list of files. The result is reflected in the tbls field.

Usage

Tool$tidy(tidy = TRUE, keep_raw = FALSE)

Arguments

tidy

(logical(1))
Should the raw parsed tibbles get tidied?

keep_raw

(logical(1))
Should the raw parsed tibbles be kept in the final output?

Returns

(R6::R6Class())
R6 object invisibly.

Method write()

Write tidy tibbles.

Usage

Tool$write(
  diro = ".",
  format = "tsv",
  input_id = NULL,
  output_id = NULL,
  pfix_include = FALSE,
  dbconn = NULL
)

Arguments

diro

(character(1))
Directory path to output tidy files. Ignored if format is db.

format

(character(1))
Format of output.

input_id

(character(1))
Input ID to use for the dataset (e.g. run123).

output_id

(character(1))
Output ID to use for the dataset (e.g. run123).

pfix_include

(logical(1))
If TRUE, prepend an input_pfix column to each tidy table.

dbconn

(DBIConnection)
Database connection object (see DBI::dbConnect).

Returns

(tibble() or NULL)
A tibble with columns raw_path, tool_parser, prefix, tidy_data, tbl_name and outpath, invisibly. NULL if no files were found.

Method nemofy()

Parse, filter, tidy and write files.

Usage

Tool$nemofy(
  diro = ".",
  format = "tsv",
  input_id = NULL,
  output_id = NULL,
  pfix_include = FALSE,
  dbconn = NULL,
  include = NULL,
  exclude = NULL
)

Arguments

diro

(character(1))
Directory path to output tidy files.

format

(character(1))
Format of output.

input_id

(character(1))
Input ID to use for the dataset (e.g. run123).

output_id

(character(1))
Output ID to use for the dataset (e.g. run123).

pfix_include

(logical(1))
If TRUE, prepend an input_pfix column to each tidy table.

dbconn

(DBIConnection)
Database connection object (see DBI::dbConnect).

include

(character(n))
tool_parser names to include (e.g. "tool1_table1").

exclude

(character(n))
tool_parser names to exclude (e.g. "tool1_table3").

Returns

(tibble() or NULL)
A tibble with columns tool_parser, prefix, tidy_data, tbl_name, outpath, invisibly. NULL if no files were found.

Method clone()

The objects of this class are cloneable with this method.

Usage

Tool$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

fs::path(tempdir(), letters[1:5]) |>
  fs::file_temp_push() |>
  fs::dir_create()
name <- "tool1"; pkg <- "nemo";
path <- system.file("extdata/tool1", package = "nemo")
toolA <- Tool$new(name = name, pkg = pkg, path = path)
toolA$files
#> # A tibble: 12 × 9
#>    tool_parser parser bname  size lastmodified        path  pattern prefix group
#>    <glue>      <chr>  <chr> <fs:> <dttm>              <chr> <chr>   <glue> <glu>
#>  1 tool1_tabl… table1 samp…    93 2026-06-04 15:05:16 /hom… "\\.to… sampl…      
#>  2 tool1_tabl… table1 samp…   113 2026-06-04 15:05:16 /hom… "\\.to… sampl… _2   
#>  3 tool1_tabl… table1 samp…   133 2026-06-04 15:05:16 /hom… "\\.to… sampl… _3   
#>  4 tool1_tabl… table2 samp…    47 2026-06-04 15:05:16 /hom… "\\.to… sampl…      
#>  5 tool1_tabl… table2 samp…    70 2026-06-04 15:05:16 /hom… "\\.to… sampl… _2   
#>  6 tool1_tabl… table3 samp…    48 2026-06-04 15:05:16 /hom… "\\.to… sampl…      
#>  7 tool1_tabl… table3 samp…    83 2026-06-04 15:05:16 /hom… "\\.to… sampl… _2   
#>  8 tool1_tabl… table4 samp…    34 2026-06-04 15:05:16 /hom… "\\.to… sampl…      
#>  9 tool1_tabl… table4 samp…    52 2026-06-04 15:05:16 /hom… "\\.to… sampl… _2   
#> 10 tool1_tabl… table5 samp…   994 2026-06-04 15:05:16 /hom… "\\.to… sampl…      
#> 11 tool1_tabl… table6 samp…    78 2026-06-04 15:05:16 /hom… "\\.to… sampl…      
#> 12 tool1_tabl… table6 samp…   100 2026-06-04 15:05:16 /hom… "\\.to… sampl… _2   
toolA$filter_files(exclude = "tool1_table3"); toolA$files # note the exclusion this time
#> # A tibble: 10 × 9
#>    tool_parser parser bname  size lastmodified        path  pattern prefix group
#>    <glue>      <chr>  <chr> <fs:> <dttm>              <chr> <chr>   <glue> <glu>
#>  1 tool1_tabl… table1 samp…    93 2026-06-04 15:05:16 /hom… "\\.to… sampl…      
#>  2 tool1_tabl… table1 samp…   113 2026-06-04 15:05:16 /hom… "\\.to… sampl… _2   
#>  3 tool1_tabl… table1 samp…   133 2026-06-04 15:05:16 /hom… "\\.to… sampl… _3   
#>  4 tool1_tabl… table2 samp…    47 2026-06-04 15:05:16 /hom… "\\.to… sampl…      
#>  5 tool1_tabl… table2 samp…    70 2026-06-04 15:05:16 /hom… "\\.to… sampl… _2   
#>  6 tool1_tabl… table4 samp…    34 2026-06-04 15:05:16 /hom… "\\.to… sampl…      
#>  7 tool1_tabl… table4 samp…    52 2026-06-04 15:05:16 /hom… "\\.to… sampl… _2   
#>  8 tool1_tabl… table5 samp…   994 2026-06-04 15:05:16 /hom… "\\.to… sampl…      
#>  9 tool1_tabl… table6 samp…    78 2026-06-04 15:05:16 /hom… "\\.to… sampl…      
#> 10 tool1_tabl… table6 samp…   100 2026-06-04 15:05:16 /hom… "\\.to… sampl… _2   
toolA$list_files()
#> # A tibble: 12 × 9
#>    tool_parser parser bname  size lastmodified        path  pattern prefix group
#>    <glue>      <chr>  <chr> <fs:> <dttm>              <chr> <chr>   <glue> <glu>
#>  1 tool1_tabl… table1 samp…    93 2026-06-04 15:05:16 /hom… "\\.to… sampl…      
#>  2 tool1_tabl… table1 samp…   113 2026-06-04 15:05:16 /hom… "\\.to… sampl… _2   
#>  3 tool1_tabl… table1 samp…   133 2026-06-04 15:05:16 /hom… "\\.to… sampl… _3   
#>  4 tool1_tabl… table2 samp…    47 2026-06-04 15:05:16 /hom… "\\.to… sampl…      
#>  5 tool1_tabl… table2 samp…    70 2026-06-04 15:05:16 /hom… "\\.to… sampl… _2   
#>  6 tool1_tabl… table3 samp…    48 2026-06-04 15:05:16 /hom… "\\.to… sampl…      
#>  7 tool1_tabl… table3 samp…    83 2026-06-04 15:05:16 /hom… "\\.to… sampl… _2   
#>  8 tool1_tabl… table4 samp…    34 2026-06-04 15:05:16 /hom… "\\.to… sampl…      
#>  9 tool1_tabl… table4 samp…    52 2026-06-04 15:05:16 /hom… "\\.to… sampl… _2   
#> 10 tool1_tabl… table5 samp…   994 2026-06-04 15:05:16 /hom… "\\.to… sampl…      
#> 11 tool1_tabl… table6 samp…    78 2026-06-04 15:05:16 /hom… "\\.to… sampl…      
#> 12 tool1_tabl… table6 samp…   100 2026-06-04 15:05:16 /hom… "\\.to… sampl… _2   

toolB <- Tool$new(name = name, pkg = pkg, path = path)$
  filter_files(include = "tool1_table1")
toolB$files
#> # A tibble: 3 × 9
#>   tool_parser  parser bname  size lastmodified        path  pattern prefix group
#>   <glue>       <chr>  <chr> <fs:> <dttm>              <chr> <chr>   <glue> <glu>
#> 1 tool1_table1 table1 samp…    93 2026-06-04 15:05:16 /hom… "\\.to… sampl…      
#> 2 tool1_table1 table1 samp…   113 2026-06-04 15:05:16 /hom… "\\.to… sampl… _2   
#> 3 tool1_table1 table1 samp…   133 2026-06-04 15:05:16 /hom… "\\.to… sampl… _3   
# tidy + write
toolC <- Tool$new(name = name, pkg = pkg, path = path)$
  filter_files(exclude = "tool1_table5")$
  tidy()
toolC$files
#> # A tibble: 11 × 9
#>    tool_parser parser bname  size lastmodified        path  pattern prefix group
#>    <glue>      <chr>  <chr> <fs:> <dttm>              <chr> <chr>   <glue> <glu>
#>  1 tool1_tabl… table1 samp…    93 2026-06-04 15:05:16 /hom… "\\.to… sampl…      
#>  2 tool1_tabl… table1 samp…   113 2026-06-04 15:05:16 /hom… "\\.to… sampl… _2   
#>  3 tool1_tabl… table1 samp…   133 2026-06-04 15:05:16 /hom… "\\.to… sampl… _3   
#>  4 tool1_tabl… table2 samp…    47 2026-06-04 15:05:16 /hom… "\\.to… sampl…      
#>  5 tool1_tabl… table2 samp…    70 2026-06-04 15:05:16 /hom… "\\.to… sampl… _2   
#>  6 tool1_tabl… table3 samp…    48 2026-06-04 15:05:16 /hom… "\\.to… sampl…      
#>  7 tool1_tabl… table3 samp…    83 2026-06-04 15:05:16 /hom… "\\.to… sampl… _2   
#>  8 tool1_tabl… table4 samp…    34 2026-06-04 15:05:16 /hom… "\\.to… sampl…      
#>  9 tool1_tabl… table4 samp…    52 2026-06-04 15:05:16 /hom… "\\.to… sampl… _2   
#> 10 tool1_tabl… table6 samp…    78 2026-06-04 15:05:16 /hom… "\\.to… sampl…      
#> 11 tool1_tabl… table6 samp…   100 2026-06-04 15:05:16 /hom… "\\.to… sampl… _2   
toolC$tbls # note the tidy column
#> # A tibble: 11 × 10
#>    tool_parser parser bname  size lastmodified        path  pattern prefix group
#>    <glue>      <chr>  <chr> <fs:> <dttm>              <chr> <chr>   <glue> <glu>
#>  1 tool1_tabl… table1 samp…    93 2026-06-04 15:05:16 /hom… "\\.to… sampl…      
#>  2 tool1_tabl… table1 samp…   113 2026-06-04 15:05:16 /hom… "\\.to… sampl… _2   
#>  3 tool1_tabl… table1 samp…   133 2026-06-04 15:05:16 /hom… "\\.to… sampl… _3   
#>  4 tool1_tabl… table2 samp…    47 2026-06-04 15:05:16 /hom… "\\.to… sampl…      
#>  5 tool1_tabl… table2 samp…    70 2026-06-04 15:05:16 /hom… "\\.to… sampl… _2   
#>  6 tool1_tabl… table3 samp…    48 2026-06-04 15:05:16 /hom… "\\.to… sampl…      
#>  7 tool1_tabl… table3 samp…    83 2026-06-04 15:05:16 /hom… "\\.to… sampl… _2   
#>  8 tool1_tabl… table4 samp…    34 2026-06-04 15:05:16 /hom… "\\.to… sampl…      
#>  9 tool1_tabl… table4 samp…    52 2026-06-04 15:05:16 /hom… "\\.to… sampl… _2   
#> 10 tool1_tabl… table6 samp…    78 2026-06-04 15:05:16 /hom… "\\.to… sampl…      
#> 11 tool1_tabl… table6 samp…   100 2026-06-04 15:05:16 /hom… "\\.to… sampl… _2   
#> # ℹ 1 more variable: tidy <list>
dir1 <- fs::file_temp(); dir2 <- fs::file_temp()
toolC$write(diro = dir1, format = "parquet", input_id = "run1")
(lfC <- list.files(dir1, full.names = TRUE))
#>  [1] "/tmp/RtmpgZV5jJ/a/sampleA_2_tool1_table1.parquet"
#>  [2] "/tmp/RtmpgZV5jJ/a/sampleA_2_tool1_table2.parquet"
#>  [3] "/tmp/RtmpgZV5jJ/a/sampleA_2_tool1_table3.parquet"
#>  [4] "/tmp/RtmpgZV5jJ/a/sampleA_2_tool1_table4.parquet"
#>  [5] "/tmp/RtmpgZV5jJ/a/sampleA_2_tool1_table6.parquet"
#>  [6] "/tmp/RtmpgZV5jJ/a/sampleA_3_tool1_table1.parquet"
#>  [7] "/tmp/RtmpgZV5jJ/a/sampleA_tool1_table1.parquet"  
#>  [8] "/tmp/RtmpgZV5jJ/a/sampleA_tool1_table2.parquet"  
#>  [9] "/tmp/RtmpgZV5jJ/a/sampleA_tool1_table3.parquet"  
#> [10] "/tmp/RtmpgZV5jJ/a/sampleA_tool1_table4.parquet"  
#> [11] "/tmp/RtmpgZV5jJ/a/sampleA_tool1_table6.parquet"  

# nemofy
toolD <- Tool$new(name = name, pkg = pkg, path = path)$
  filter_files(exclude = "tool1_table5")$
  nemofy(diro = dir2, format = "parquet", input_id = "run2")
(lfD <- list.files(dir2, full.names = TRUE))
#>  [1] "/tmp/RtmpgZV5jJ/b/sampleA_2_tool1_table1.parquet"
#>  [2] "/tmp/RtmpgZV5jJ/b/sampleA_2_tool1_table2.parquet"
#>  [3] "/tmp/RtmpgZV5jJ/b/sampleA_2_tool1_table3.parquet"
#>  [4] "/tmp/RtmpgZV5jJ/b/sampleA_2_tool1_table4.parquet"
#>  [5] "/tmp/RtmpgZV5jJ/b/sampleA_2_tool1_table6.parquet"
#>  [6] "/tmp/RtmpgZV5jJ/b/sampleA_3_tool1_table1.parquet"
#>  [7] "/tmp/RtmpgZV5jJ/b/sampleA_tool1_table1.parquet"  
#>  [8] "/tmp/RtmpgZV5jJ/b/sampleA_tool1_table2.parquet"  
#>  [9] "/tmp/RtmpgZV5jJ/b/sampleA_tool1_table3.parquet"  
#> [10] "/tmp/RtmpgZV5jJ/b/sampleA_tool1_table4.parquet"  
#> [11] "/tmp/RtmpgZV5jJ/b/sampleA_tool1_table6.parquet"