Workflow Object

Orchestrates multiple Tool objects over a shared directory, handling file discovery, filtering, tidying, and writing in a single pipeline.

A Workflow object:

has a name (name);

has one or more paths to workflow results (path);

has a list of Tool subclass objects (tools);

has a tibble of all files in the shared directory (files_tbl);

has a tibble of written output files, populated after write() (written_files);

The typical workflow is: optionally filter with filter_files(), tidy with tidy(), then write with write(). nemofy() chains all three steps.

Public fields

name

(character(1))
Name of workflow.

path

(character(n))
Path(s) to workflow results.

tools

(list(n))
List of Tools that compose a Workflow.

files_tbl

(tibble(n))
Tibble of files from list_files_dir().

metapkg

(character(1))
Package name used for metadata version reporting.

written_files

(tibble(n))
Tibble of files written from self$write().

Methods

Public methods

Workflow$new()

Workflow$print()

Workflow$filter_files()

Workflow$list_files()

Workflow$tidy()

Workflow$write()

Workflow$nemofy()

Workflow$get_raw_schemas_all()

Workflow$get_tidy_schemas_all()

Workflow$get_tbls()

Workflow$get_metadata()

Workflow$clone()

Method new()

Create a new Workflow object.

Usage

Workflow$new(name = NULL, path = NULL, tools = NULL, metapkg = "nemo")

Arguments

name

(character(1))
Name of workflow.

path

(character(n))
Path(s) to workflow results.

tools

(list(n))
List of Tools that compose a Workflow.

metapkg

(character(1))
Package name used for metadata version reporting.

Returns

(R6::R6Class())
R6 object.

Method print()

Print details about the Workflow.

Usage

Workflow$print(...)

Arguments

...

(ignored).

Returns

(R6::R6Class())
R6 object invisibly.

Method filter_files()

Filter files in given workflow directory.

Usage

Workflow$filter_files(include = NULL, exclude = NULL)

Arguments

include

(character(n))
tool_parser names to include (e.g. "tool1_table1").

exclude

(character(n))
tool_parser names to exclude (e.g. "tool1_table5").

Returns

(R6::R6Class())
R6 object invisibly.

Method list_files()

List only files of interest in given workflow directory, i.e. only those files that match the patterns listed in the individual tool config.

Usage

Workflow$list_files(type = c("file", "symlink"))

Arguments

type

(character(1))
File types(s) to return (e.g. any, file, directory, symlink). See fs::dir_info.

Returns

(tibble())
Bound list_files() tibbles from all Tools.

Method tidy()

Tidy Workflow files.

Usage

Workflow$tidy(tidy = TRUE, keep_raw = FALSE)

Arguments

tidy

(logical(1))
Should the raw parsed tibbles get tidied?

keep_raw

(logical(1))
Should the raw parsed tibbles be kept in the final output?

Returns

(R6::R6Class())
R6 object invisibly.

Method write()

Write tidy tibbles.

Usage

Workflow$write(
  diro = ".",
  format = "tsv",
  input_id = NULL,
  output_id = NULL,
  pfix_include = FALSE,
  dbconn = NULL
)

Arguments

diro

(character(1))
Directory path to output tidy files.

format

(character(1))
Format of output.

input_id

(character(1))
Input ID to use for the dataset (e.g. run123).

output_id

(character(1))
Output ID to use for the dataset (e.g. run123).

pfix_include

(logical(1))
If TRUE, prepend an input_pfix column to each tidy table.

dbconn

(DBIConnection)
Database connection object (see DBI::dbConnect).

Returns

(R6::R6Class())
R6 object invisibly.

Method nemofy()

Parse, filter, tidy and write files.

Usage

Workflow$nemofy(
  diro = ".",
  format = "tsv",
  input_id = NULL,
  output_id = NULL,
  pfix_include = FALSE,
  dbconn = NULL,
  include = NULL,
  exclude = NULL
)

Arguments

diro

(character(1))
Directory path to output tidy files.

format

(character(1))
Format of output.

input_id

(character(1))
Input ID to use for the dataset (e.g. run123).

output_id

(character(1))
Output ID to use for the dataset (e.g. run123).

pfix_include

(logical(1))
If TRUE, prepend an input_pfix column to each tidy table.

dbconn

(DBIConnection)
Database connection object (see DBI::dbConnect).

include

(character(n))
tool_parser names to include (e.g. "tool1_table1").

exclude

(character(n))
tool_parser names to exclude (e.g. "tool1_table5").

Returns

(R6::R6Class())
R6 object invisibly.

Method get_raw_schemas_all()

Get raw schemas for all Tools.

Usage

Workflow$get_raw_schemas_all()

Returns

(tibble())
Bound raw_schemas_all tibbles from all Tools, with a leading tool column.

Method get_tidy_schemas_all()

Get tidy schemas for all Tools.

Usage

Workflow$get_tidy_schemas_all()

Returns

(tibble())
Bound tidy schema tibbles from all Tools, with a leading tool column.

Method get_tbls()

Get tidy tibbles for all Tools.

Usage

Workflow$get_tbls()

Returns

(tibble())
Bound tbls tibbles from all Tools.

Method get_metadata()

Get metadata for the workflow run.

Usage

Workflow$get_metadata(input_id, output_id, output_dir, pkgs = NULL)

Arguments

input_id

(character(1))
Input ID to use for the dataset (e.g. run123).

output_id

(character(1))
Output ID to use for the dataset (e.g. run123).

output_dir

(character(1))
Output directory.

pkgs

(character(n))
Which R packages to extract versions for.

Returns

(list())
List with input_id, output_id, input_dirs, output_dir, pkg_versions, and files.

Method clone()

The objects of this class are cloneable with this method.

Usage

Workflow$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

fs::path(tempdir(), letters[1:5]) |>
  fs::file_temp_push() |>
  fs::dir_create()
path <- system.file("extdata/tool1", package = "nemo")
tools <- list(tool1 = Tool1)
wf <- Workflow$new(name = "wf1", path = path, tools = tools)
(lf_all <- wf$list_files())
#> # A tibble: 12 × 9
#>    tool_parser parser bname  size lastmodified        path  pattern prefix group
#>    <glue>      <chr>  <chr> <fs:> <dttm>              <chr> <chr>   <glue> <glu>
#>  1 tool1_tabl… table1 samp…    93 2026-05-15 02:17:48 /hom… "\\.to… sampl…      
#>  2 tool1_tabl… table1 samp…   113 2026-05-15 02:17:48 /hom… "\\.to… sampl… _2   
#>  3 tool1_tabl… table1 samp…   133 2026-05-15 02:17:48 /hom… "\\.to… sampl… _3   
#>  4 tool1_tabl… table2 samp…    47 2026-05-15 02:17:48 /hom… "\\.to… sampl…      
#>  5 tool1_tabl… table2 samp…    70 2026-05-15 02:17:48 /hom… "\\.to… sampl… _2   
#>  6 tool1_tabl… table3 samp…    48 2026-05-15 02:17:48 /hom… "\\.to… sampl…      
#>  7 tool1_tabl… table3 samp…    83 2026-05-15 02:17:48 /hom… "\\.to… sampl… _2   
#>  8 tool1_tabl… table4 samp…    34 2026-05-15 02:17:48 /hom… "\\.to… sampl…      
#>  9 tool1_tabl… table4 samp…    52 2026-05-15 02:17:48 /hom… "\\.to… sampl… _2   
#> 10 tool1_tabl… table5 samp…   994 2026-05-15 02:17:48 /hom… "\\.to… sampl…      
#> 11 tool1_tabl… table6 samp…    78 2026-05-15 02:17:48 /hom… "\\.to… sampl…      
#> 12 tool1_tabl… table6 samp…   100 2026-05-15 02:17:48 /hom… "\\.to… sampl… _2   
wf$filter_files(exclude = "tool1_table5")
wf$tidy()
(tbls <- wf$get_tbls())
#> # A tibble: 11 × 10
#>    tool_parser parser bname  size lastmodified        path  pattern prefix group
#>    <glue>      <chr>  <chr> <fs:> <dttm>              <chr> <chr>   <glue> <glu>
#>  1 tool1_tabl… table1 samp…    93 2026-05-15 02:17:48 /hom… "\\.to… sampl…      
#>  2 tool1_tabl… table1 samp…   113 2026-05-15 02:17:48 /hom… "\\.to… sampl… _2   
#>  3 tool1_tabl… table1 samp…   133 2026-05-15 02:17:48 /hom… "\\.to… sampl… _3   
#>  4 tool1_tabl… table2 samp…    47 2026-05-15 02:17:48 /hom… "\\.to… sampl…      
#>  5 tool1_tabl… table2 samp…    70 2026-05-15 02:17:48 /hom… "\\.to… sampl… _2   
#>  6 tool1_tabl… table3 samp…    48 2026-05-15 02:17:48 /hom… "\\.to… sampl…      
#>  7 tool1_tabl… table3 samp…    83 2026-05-15 02:17:48 /hom… "\\.to… sampl… _2   
#>  8 tool1_tabl… table4 samp…    34 2026-05-15 02:17:48 /hom… "\\.to… sampl…      
#>  9 tool1_tabl… table4 samp…    52 2026-05-15 02:17:48 /hom… "\\.to… sampl… _2   
#> 10 tool1_tabl… table6 samp…    78 2026-05-15 02:17:48 /hom… "\\.to… sampl…      
#> 11 tool1_tabl… table6 samp…   100 2026-05-15 02:17:48 /hom… "\\.to… sampl… _2   
#> # ℹ 1 more variable: tidy <list>
(rs <- wf$get_raw_schemas_all())
#> # A tibble: 12 × 5
#>    tool  name   tbl_description                                 version schema  
#>    <chr> <chr>  <chr>                                           <chr>   <list>  
#>  1 tool1 table1 Table1 for tool1 (txt: header present, tab-del… v1.2.3  <tibble>
#>  2 tool1 table1 Table1 for tool1 (txt: header present, tab-del… v4.5.6  <tibble>
#>  3 tool1 table1 Table1 for tool1 (txt: header present, tab-del… latest  <tibble>
#>  4 tool1 table2 Table2 for tool1 (txt: header present, tab-del… v1.0.0  <tibble>
#>  5 tool1 table2 Table2 for tool1 (txt: header present, tab-del… latest  <tibble>
#>  6 tool1 table3 Table3 for tool1 (txt-keyvalue: no header, 2 c… v1.0.0  <tibble>
#>  7 tool1 table3 Table3 for tool1 (txt-keyvalue: no header, 2 c… latest  <tibble>
#>  8 tool1 table4 Table4 for tool1 (txt-nohead: no header, posit… v1.0.0  <tibble>
#>  9 tool1 table4 Table4 for tool1 (txt-nohead: no header, posit… latest  <tibble>
#> 10 tool1 table6 Table6 for tool1 (csv: header present, comma-d… v1.0.0  <tibble>
#> 11 tool1 table6 Table6 for tool1 (csv: header present, comma-d… latest  <tibble>
#> 12 tool1 table5 Table5 for tool1 (csv-nohead-long: no header, … latest  <tibble>
dir1 <- fs::file_temp(); dir2 <- fs::file_temp()
wf$write(diro = dir1, format = "parquet", input_id = "run1")
(lf1 <- list.files(dir1, pattern = "tool1.*parquet", full.names = TRUE))
#>  [1] "/tmp/RtmpX1i0hJ/c/sampleA_2_tool1_table1.parquet"
#>  [2] "/tmp/RtmpX1i0hJ/c/sampleA_2_tool1_table2.parquet"
#>  [3] "/tmp/RtmpX1i0hJ/c/sampleA_2_tool1_table3.parquet"
#>  [4] "/tmp/RtmpX1i0hJ/c/sampleA_2_tool1_table4.parquet"
#>  [5] "/tmp/RtmpX1i0hJ/c/sampleA_2_tool1_table6.parquet"
#>  [6] "/tmp/RtmpX1i0hJ/c/sampleA_3_tool1_table1.parquet"
#>  [7] "/tmp/RtmpX1i0hJ/c/sampleA_tool1_table1.parquet"  
#>  [8] "/tmp/RtmpX1i0hJ/c/sampleA_tool1_table2.parquet"  
#>  [9] "/tmp/RtmpX1i0hJ/c/sampleA_tool1_table3.parquet"  
#> [10] "/tmp/RtmpX1i0hJ/c/sampleA_tool1_table4.parquet"  
#> [11] "/tmp/RtmpX1i0hJ/c/sampleA_tool1_table6.parquet"  
(meta <- wf$get_metadata(input_id = "run1", output_id = "out1", output_dir = dir1))
#> $input_id
#> [x] "run1"
#> 
#> $output_id
#> [x] "out1"
#> 
#> $input_dirs
#> [1] "/home/runner/miniconda3/envs/pkgdown_env/lib/R/library/nemo/extdata/tool1"
#> 
#> $output_dir
#> [x] "/tmp/RtmpX1i0hJ/c"
#> 
#> $pkg_versions
#> # A tibble: 1 × 2
#>   name  version   
#>   <chr> <chr>     
#> 1 nemo  0.0.3.9016
#> 
#> $files
#> # A tibble: 11 × 4
#>    tbl          prefix    fout                           fin                    
#>    <glue>       <glue>    <chr>                          <chr>                  
#>  1 tool1_table1 sampleA   sampleA_tool1_table1.parquet   /home/runner/miniconda…
#>  2 tool1_table1 sampleA_2 sampleA_2_tool1_table1.parquet /home/runner/miniconda…
#>  3 tool1_table1 sampleA_3 sampleA_3_tool1_table1.parquet /home/runner/miniconda…
#>  4 tool1_table2 sampleA   sampleA_tool1_table2.parquet   /home/runner/miniconda…
#>  5 tool1_table2 sampleA_2 sampleA_2_tool1_table2.parquet /home/runner/miniconda…
#>  6 tool1_table3 sampleA   sampleA_tool1_table3.parquet   /home/runner/miniconda…
#>  7 tool1_table3 sampleA_2 sampleA_2_tool1_table3.parquet /home/runner/miniconda…
#>  8 tool1_table4 sampleA   sampleA_tool1_table4.parquet   /home/runner/miniconda…
#>  9 tool1_table4 sampleA_2 sampleA_2_tool1_table4.parquet /home/runner/miniconda…
#> 10 tool1_table6 sampleA   sampleA_tool1_table6.parquet   /home/runner/miniconda…
#> 11 tool1_table6 sampleA_2 sampleA_2_tool1_table6.parquet /home/runner/miniconda…
#> 
wf2 <- Workflow$new(name = "wf2", path = path, tools = tools)
wf2$nemofy(diro = dir2, format = "parquet", input_id = "run2")
(lf2 <- list.files(dir2, pattern = "tool1.*parquet", full.names = TRUE))
#>  [1] "/tmp/RtmpX1i0hJ/d/sampleA_2_tool1_table1.parquet"
#>  [2] "/tmp/RtmpX1i0hJ/d/sampleA_2_tool1_table2.parquet"
#>  [3] "/tmp/RtmpX1i0hJ/d/sampleA_2_tool1_table3.parquet"
#>  [4] "/tmp/RtmpX1i0hJ/d/sampleA_2_tool1_table4.parquet"
#>  [5] "/tmp/RtmpX1i0hJ/d/sampleA_2_tool1_table6.parquet"
#>  [6] "/tmp/RtmpX1i0hJ/d/sampleA_3_tool1_table1.parquet"
#>  [7] "/tmp/RtmpX1i0hJ/d/sampleA_tool1_table1.parquet"  
#>  [8] "/tmp/RtmpX1i0hJ/d/sampleA_tool1_table2.parquet"  
#>  [9] "/tmp/RtmpX1i0hJ/d/sampleA_tool1_table3.parquet"  
#> [10] "/tmp/RtmpX1i0hJ/d/sampleA_tool1_table4.parquet"  
#> [11] "/tmp/RtmpX1i0hJ/d/sampleA_tool1_table5.parquet"  
#> [12] "/tmp/RtmpX1i0hJ/d/sampleA_tool1_table6.parquet"